+++ appcontext/context.go
@@ -2,25 +2,27 @@ package appcontext
 
 import "sync"
 
-//List of consts containing the names of the available componentes in the Application Context - appcontext.Current
+// List of consts containing the names of the available componentes in the Application Context - appcontext.Current
 const (
-	Spell    = "Spell"
-	Renderer = "Renderer"
+	AddComponentConstantToContext = "AddComponentConstantToContext"
+	AddPackageImportToMain        = "AddPackageImportToMain"
+	Spell                         = "Spell"
+	Renderer                      = "Renderer"
 )
 
-//Component is the Base interface for all Components
+// Component is the Base interface for all Components
 type Component interface{}
 
-//ComponentInitializerFunction specifies a function for lazily initializing a component
+// ComponentInitializerFunction specifies a function for lazily initializing a component
 type ComponentInitializerFunction func() Component
 
-//ComponentInfo holds the function to lazy initialize the component and the instance created following the singleton pattern
+// ComponentInfo holds the function to lazy initialize the component and the instance created following the singleton pattern
 type ComponentInfo struct {
 	Initializer ComponentInitializerFunction
 	Instance    Component
 }
 
-//Get s the instance. If it is not created, creates and stores it to the next calls
+// Get s the instance. If it is not created, creates and stores it to the next calls
 func (componentInfo *ComponentInfo) Get() Component {
 	if componentInfo.Instance == nil {
 		componentInfo.Instance = componentInfo.Initializer()
@@ -29,16 +31,16 @@ func (componentInfo *ComponentInfo) Get() Component {
 	return componentInfo.Instance
 }
 
-//ApplicationContext is the type defining a map of Components
+// ApplicationContext is the type defining a map of Components
 type ApplicationContext struct {
 	components     map[string]*ComponentInfo
 	componentMutex sync.Mutex
 }
 
-//Current keeps all components available, initialized in the application startup
+// Current keeps all components available, initialized in the application startup
 var Current ApplicationContext
 
-//Add a component By Name
+// Add a component By Name
 func (applicationContext *ApplicationContext) Add(componentName string, componentInitializerFunction ComponentInitializerFunction) {
 	applicationContext.componentMutex.Lock()
 	defer applicationContext.componentMutex.Unlock()
@@ -46,7 +48,7 @@ func (applicationContext *ApplicationContext) Add(componentName string, componen
 	applicationContext.components[componentName] = &ComponentInfo{Initializer: componentInitializerFunction}
 }
 
-//Get a component By Name
+// Get a component By Name
 func (applicationContext *ApplicationContext) Get(componentName string) Component {
 	if applicationContext.components[componentName] == nil {
 		return nil
@@ -54,7 +56,7 @@ func (applicationContext *ApplicationContext) Get(componentName string) Componen
 	return applicationContext.components[componentName].Get()
 }
 
-//Delete a component By Name
+// Delete a component By Name
 func (applicationContext *ApplicationContext) Delete(componentName string) {
 	applicationContext.componentMutex.Lock()
 	defer applicationContext.componentMutex.Unlock()
@@ -62,7 +64,7 @@ func (applicationContext *ApplicationContext) Delete(componentName string) {
 	delete(applicationContext.components, componentName)
 }
 
-//Count returns the count of components registered
+// Count returns the count of components registered
 func (applicationContext *ApplicationContext) Count() int {
 	applicationContext.componentMutex.Lock()
 	defer applicationContext.componentMutex.Unlock()
@@ -70,7 +72,7 @@ func (applicationContext *ApplicationContext) Count() int {
 	return len(applicationContext.components)
 }
 
-//CreateApplicationContext creates a new ApplicationContext instance
+// CreateApplicationContext creates a new ApplicationContext instance
 func CreateApplicationContext() ApplicationContext {
 	return ApplicationContext{components: make(map[string]*ComponentInfo)}
 }